refactor: Replace old genetic stats with new
